// frontend/src/Models/Ad.ts
export type Produto =  {
  nome: string;
  DataPublicacao: Date;
  valor: number;
  cidade: string;
  categoria: string;
  modelo: string;
  condicao: string;
  quantidade: string;
}

export type Servico = {
  nome: string;
  DataPublicacao: Date;
  valor: number;
  cidade: string;
}